fix: rename muscle-groups.ts to fix Convex build
Convex only allows alphanumeric characters, underscores, and periods
in file paths. The hyphen in 'muscle-groups.ts' caused build failures.

Renamed: convex/lib/muscle-groups.ts → convex/lib/muscleGroups.ts
