+++ convex/_generated/api.d.ts
@@ -25,6 +25,7 @@ import type * as analyticsProgressiveOverload from "../analyticsProgressiveOverl
 import type * as analyticsRecovery from "../analyticsRecovery.js";
 import type * as crons from "../crons.js";
 import type * as exercises from "../exercises.js";
+import type * as lib_muscleGroups from "../lib/muscleGroups.js";
 import type * as lib_pr_detection from "../lib/pr_detection.js";
 import type * as lib_rateLimit from "../lib/rateLimit.js";
 import type * as lib_streak_calculator from "../lib/streak_calculator.js";
@@ -58,6 +59,7 @@ declare const fullApi: ApiFromModules<{
   analyticsRecovery: typeof analyticsRecovery;
   crons: typeof crons;
   exercises: typeof exercises;
+  "lib/muscleGroups": typeof lib_muscleGroups;
   "lib/pr_detection": typeof lib_pr_detection;
   "lib/rateLimit": typeof lib_rateLimit;
   "lib/streak_calculator": typeof lib_streak_calculator;
